requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.23.0
pandas==2.2.2
numpy==1.26.0
matplotlib==3.8.1
scikit-learn==1.3.1
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
pillow==10.1.0
pytest==7.4.0
duckdb==1.3.2
lxml>=5.3.0
